docker/kim-iot-client-base-env/src/requirements.txt
pytest>=3.0.7
logzero  # see github.com/metachris/logzero
pyaudio
pyzmq

